Clean up merge conflict of PetService, PetServiceImpl and PetRepository
