+++ backend/src/main/java/com/remitk/registry/repository/PetRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface PetRepository extends JpaRepository<Pet, Long> {
     Page<Pet> findByNameContainingIgnoreCase(String name, Pageable pageable);
     Page<Pet> findByOwnerId(Long id, Pageable pageable);
-    Page<Pet> findByMicrochipId(Long id, Pageable pageable);
+    Optional<Pet> findByMicrochipId(Long id);
 }
